# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PGE-QROO",
        "description": "Documentación de la API para la Plataforma de Gestión Energética de Quintana Roo. Permite la gestión de dependencias, edificios, presupuestos y análisis de consumo.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dependencias/{dependencia}/edificios": {
            "get": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Listar edificios de una dependencia",
                "description": "Devuelve todos los edificios asociados a una dependencia específica.",
                "operationId": "23968fcfb0a02d21f0e8166a34064302",
                "parameters": [
                    {
                        "name": "dependencia",
                        "in": "path",
                        "description": "ID de la dependencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de edificios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_edificio": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "nombre_edificio": {
                                                "type": "string",
                                                "example": "Edificio Central"
                                            },
                                            "direccion": {
                                                "type": "string",
                                                "example": "Av. Principal 123"
                                            },
                                            "latitud": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 18.5001
                                            },
                                            "longitud": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -88.3002
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Dependencia no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Crear edificio",
                "description": "Crea un nuevo edificio dentro de una dependencia.",
                "operationId": "e38fd71e63c33d05e6f0bf01c0c5843e",
                "parameters": [
                    {
                        "name": "dependencia",
                        "in": "path",
                        "description": "ID de la dependencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_edificio"
                                ],
                                "properties": {
                                    "nombre_edificio": {
                                        "type": "string",
                                        "example": "Edificio Norte"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle 10 #45"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 19.4326
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -99.1332
                                    },
                                    "caracteristicas": {
                                        "type": "string",
                                        "example": "3 niveles, oficinas administrativas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Edificio creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_edificio": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "nombre_edificio": {
                                            "type": "string",
                                            "example": "Edificio Norte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/edificios/{edificio}": {
            "put": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Actualizar edificio",
                "description": "Actualiza los datos de un edificio existente.",
                "operationId": "0196bb80a5992d9c504844fbdaf2c370",
                "parameters": [
                    {
                        "name": "edificio",
                        "in": "path",
                        "description": "ID del edificio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_edificio": {
                                        "type": "string",
                                        "example": "Edificio Renovado"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Reforma 100"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "caracteristicas": {
                                        "type": "string",
                                        "example": "Paneles solares instalados"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Edificio actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Edificio no encontrado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Eliminar edificio",
                "description": "Elimina un edificio del sistema.",
                "operationId": "eab50c6820f063eb9cd520f8cabadc8f",
                "parameters": [
                    {
                        "name": "edificio",
                        "in": "path",
                        "description": "ID del edificio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edificio eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Edificio eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Edificio no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/integracion/gobierno": {
            "get": {
                "tags": [
                    "Integración"
                ],
                "summary": "Obtener datos oficiales del Gobierno",
                "description": "Consulta el servicio externo Núcleo Digital.",
                "operationId": "e9a3a3b22322f3491c09a784f031f926",
                "responses": {
                    "200": {
                        "description": "Datos obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Error al conectar con el Núcleo Digital"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}